+++ internal/k8s/reconciler/gateway.go
@@ -3,13 +3,10 @@ package reconciler
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"strings"
 	"sync/atomic"
 
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
 	gw "sigs.k8s.io/gateway-api/apis/v1alpha2"
 
 	"github.com/hashicorp/consul-api-gateway/internal/core"
@@ -22,10 +19,10 @@ import (
 )
 
 type K8sGateway struct {
-	GatewayState
+	*GatewayState
 	*gw.Gateway
 
-	listeners map[string]*K8sListener
+	listeners []*K8sListener
 
 	consulNamespace   string
 	logger            hclog.Logger
@@ -37,241 +34,11 @@ type K8sGateway struct {
 
 var _ store.StatusTrackingGateway = &K8sGateway{}
 
-func (g *K8sGateway) Validate(ctx context.Context) error {
-	g.GatewayState.Status = GatewayStatus{}
-	g.validateListenerConflicts()
-
-	if err := g.validatePods(ctx); err != nil {
-		return err
-	}
-
-	if err := g.validateGatewayIP(ctx); err != nil {
-		return err
-	}
-
-	for _, listener := range g.listeners {
-		if err := listener.Validate(ctx, g.client, g.Gateway, listener.listener); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-type mergedListener struct {
-	port      gw.PortNumber
-	listeners []*K8sListener
-	protocols map[string]struct{}
-	hostnames map[string]struct{}
-}
-
-func (g *K8sGateway) mergeListenersByPort() map[gw.PortNumber]mergedListener {
-	mergedListeners := make(map[gw.PortNumber]mergedListener)
-	for _, listener := range g.listeners {
-		merged, found := mergedListeners[listener.listener.Port]
-		if !found {
-			merged = mergedListener{
-				port:      listener.listener.Port,
-				protocols: make(map[string]struct{}),
-				hostnames: make(map[string]struct{}),
-			}
-		}
-		merged.listeners = append(merged.listeners, listener)
-		merged.protocols[string(listener.listener.Protocol)] = struct{}{}
-		if listener.listener.Hostname != nil {
-			merged.hostnames[string(*listener.listener.Hostname)] = struct{}{}
-		}
-		mergedListeners[listener.listener.Port] = merged
-	}
-	return mergedListeners
-}
-
-func (g *K8sGateway) validateListenerConflicts() {
-	for _, merged := range g.mergeListenersByPort() {
-		if len(merged.protocols) > 1 {
-			conflict := fmt.Errorf("listeners have conflicting protocols for port: %s", setToCSV(merged.protocols))
-			for _, listener := range merged.listeners {
-				listener.ListenerState.Status.Conflicted.ProtocolConflict = conflict
-			}
-		}
-		if len(merged.hostnames) > 1 {
-			conflict := fmt.Errorf("listeners have conflicting hostnames for port: %s", setToCSV(merged.protocols))
-			for _, listener := range merged.listeners {
-				listener.ListenerState.Status.Conflicted.HostnameConflict = conflict
-			}
-		}
-	}
-}
-
-// validateGatewayIP ensures that the appropriate IP addresses are assigned to the
-// Gateway.
-func (g *K8sGateway) validateGatewayIP(ctx context.Context) error {
-	service := g.serviceBuilder.Build()
-	if service == nil {
-		return g.assignGatewayIPFromPod(ctx)
-	}
-
-	switch service.Spec.Type {
-	case corev1.ServiceTypeLoadBalancer:
-		return g.assignGatewayIPFromServiceIngress(ctx, service)
-	case corev1.ServiceTypeClusterIP:
-		return g.assignGatewayIPFromService(ctx, service)
-	case corev1.ServiceTypeNodePort:
-		/* For serviceType: NodePort, there isn't a consistent way to guarantee access to the
-		 * service from outside the k8s cluster. For now, we're putting the IP address of the
-		 * nodes that the gateway pods are running on.
-		 * The practitioner will have to understand that they may need to port forward into the
-		 * cluster (in the case of Kind) or open firewall rules (in the case of GKE) in order to
-		 * access the gateway from outside the cluster.
-		 */
-		return g.assignGatewayIPFromPodHost(ctx)
-	default:
-		return fmt.Errorf("unsupported service type: %s", service.Spec.Type)
-	}
-}
-
-// assignGatewayIPFromServiceIngress retrieves the external load balancer
-// ingress IP for the Service and assigns it to the Gateway
-func (g *K8sGateway) assignGatewayIPFromServiceIngress(ctx context.Context, service *corev1.Service) error {
-	updated, err := g.client.GetService(ctx, types.NamespacedName{Namespace: service.Namespace, Name: service.Name})
-	if err != nil {
-		return err
-	}
-
-	if updated == nil {
-		g.GatewayState.Status.Scheduled.NotReconciled = errors.New("service not found")
-		return nil
-	}
-
-	for _, ingress := range updated.Status.LoadBalancer.Ingress {
-		g.ServiceReady = true
-		g.Addresses = append(g.Addresses, ingress.IP)
-	}
-
-	return nil
-}
-
-// assignGatewayIPFromService retrieves the internal cluster IP for the
-// Service and assigns it to the Gateway
-func (g *K8sGateway) assignGatewayIPFromService(ctx context.Context, service *corev1.Service) error {
-	updated, err := g.client.GetService(ctx, types.NamespacedName{Namespace: service.Namespace, Name: service.Name})
-	if err != nil {
-		return err
-	}
-
-	if updated == nil {
-		g.GatewayState.Status.Scheduled.NotReconciled = errors.New("service not found")
-		return nil
-	}
-
-	if updated.Spec.ClusterIP != "" {
-		g.ServiceReady = true
-		g.Addresses = append(g.Addresses, updated.Spec.ClusterIP)
-	}
-
-	return nil
-}
-
-// assignGatewayIPFromPod retrieves the internal IP for the Pod and assigns
-// it to the Gateway.
-func (g *K8sGateway) assignGatewayIPFromPod(ctx context.Context) error {
-	pod, err := g.client.PodWithLabels(ctx, utils.LabelsForGateway(g.Gateway))
-	if err != nil {
-		return err
-	}
-
-	if pod == nil {
-		g.GatewayState.Status.Scheduled.NotReconciled = errors.New("pod not found")
-		return nil
-	}
-
-	if pod.Status.PodIP != "" {
-		g.ServiceReady = true
-		g.Addresses = append(g.Addresses, pod.Status.PodIP)
-	}
-
-	return nil
-}
-
-// assignGatewayIPFromPodHost retrieves the (potentially) externally accessible
-// IP address for the host that the Pod is running on and assigns it to the Gateway.
-// This IP address is not always externally accessible and may require additional
-// work by the practitioner such as port-forwarding or opening firewall rules to make
-// it externally accessible.
-func (g *K8sGateway) assignGatewayIPFromPodHost(ctx context.Context) error {
-	pod, err := g.client.PodWithLabels(ctx, utils.LabelsForGateway(g.Gateway))
-	if err != nil {
-		return err
-	}
-
-	if pod == nil {
-		g.GatewayState.Status.Scheduled.NotReconciled = errors.New("pod not found")
-		return nil
-	}
-
-	if pod.Status.HostIP != "" {
-		g.ServiceReady = true
-		g.Addresses = append(g.Addresses, pod.Status.HostIP)
-	}
-
-	return nil
-}
-
-func (g *K8sGateway) validatePods(ctx context.Context) error {
-	pod, err := g.client.PodWithLabels(ctx, utils.LabelsForGateway(g.Gateway))
-	if err != nil {
-		return err
-	}
-
-	g.validatePodConditions(pod)
-
-	return nil
-}
-
-func (g *K8sGateway) validatePodConditions(pod *corev1.Pod) {
-	if pod == nil {
-		g.GatewayState.Status.Scheduled.NotReconciled = errors.New("pod not found")
-		return
-	}
-
-	switch pod.Status.Phase {
-	case corev1.PodPending:
-		g.validatePodStatusPending(pod)
-	case corev1.PodRunning:
-		g.validatePodStatusRunning(pod)
-	case corev1.PodSucceeded:
-		// this should never happen, occurs when the pod terminates
-		// with a 0 status code, consider this a failed deployment
-		fallthrough
-	case corev1.PodFailed:
-		// we have a failed deployment, set the status accordingly
-		// for now we just consider the pods unschedulable.
-		g.GatewayState.Status.Scheduled.PodFailed = errors.New("pod not running")
-	default: // Unknown pod status
-		// we don't have a known pod status, just consider this unreconciled
-		g.GatewayState.Status.Scheduled.Unknown = errors.New("pod status unknown")
-	}
-}
-
-func (g *K8sGateway) validatePodStatusPending(pod *corev1.Pod) {
-	for _, condition := range pod.Status.Conditions {
-		if condition.Type == corev1.PodScheduled && condition.Status == corev1.ConditionFalse &&
-			strings.Contains(condition.Reason, "Unschedulable") {
-			g.GatewayState.Status.Scheduled.NoResources = errors.New(condition.Message)
-			return
-		}
-	}
-	// if no conditions exist, or we haven't found a specific above condition, just default
-	// to not reconciled
-	g.GatewayState.Status.Scheduled.NotReconciled = errors.New("pod conditions not found")
-}
-
-func (g *K8sGateway) validatePodStatusRunning(pod *corev1.Pod) {
-	for _, condition := range pod.Status.Conditions {
-		if condition.Type == corev1.PodReady && condition.Status == corev1.ConditionTrue {
-			g.PodReady = true
-			return
-		}
+// TODO: remove this
+func (g *K8sGateway) SetState(state *GatewayState) {
+	g.GatewayState = state
+	for i, listener := range g.listeners {
+		listener.ListenerState = state.Listeners[i]
 	}
 }
 
@@ -298,23 +65,6 @@ func (g *K8sGateway) Listeners() []store.Listener {
 	}
 
 	return listeners
-}
-
-func (g *K8sGateway) ShouldUpdate(other store.Gateway) bool {
-	if other == nil {
-		return false
-	}
-
-	if g == nil {
-		return true
-	}
-
-	otherGateway, ok := other.(*K8sGateway)
-	if !ok {
-		return false
-	}
-
-	return !utils.ResourceVersionGreater(g.Gateway.ResourceVersion, otherGateway.Gateway.ResourceVersion)
 }
 
 // Bind returns the name of the listeners to which a route bound
@@ -346,53 +96,6 @@ func (g *K8sGateway) Bind(ctx context.Context, route store.Route) []string {
 	return nil
 }
 
-func (g *K8sGateway) Status() gw.GatewayStatus {
-	listenerStatuses := []gw.ListenerStatus{}
-	listenersReady := true
-	listenersInvalid := false
-	for _, listener := range g.listeners {
-		if listener.ListenerState.Status.Ready.Pending != nil {
-			listenersReady = false
-		}
-		if listener.ListenerState.Status.Ready.Invalid != nil {
-			listenersInvalid = true
-		}
-		listenerStatuses = append(listenerStatuses, listener.Status())
-	}
-
-	if listenersInvalid {
-		g.GatewayState.Status.Ready.ListenersNotValid = errors.New("gateway listeners not valid")
-	} else if !g.PodReady || !g.ServiceReady || !listenersReady {
-		g.GatewayState.Status.Ready.ListenersNotReady = errors.New("gateway listeners not ready")
-	} else if len(g.Gateway.Spec.Addresses) != 0 {
-		g.GatewayState.Status.Ready.AddressNotAssigned = errors.New("gateway does not support requesting addresses")
-	}
-	conditions := g.GatewayState.Status.Conditions(g.Gateway.Generation)
-
-	// prefer to not update to not mess up timestamps
-	if listenerStatusesEqual(listenerStatuses, g.Gateway.Status.Listeners) {
-		listenerStatuses = g.Gateway.Status.Listeners
-	}
-	if conditionsEqual(conditions, g.Gateway.Status.Conditions) {
-		conditions = g.Gateway.Status.Conditions
-	}
-
-	ipType := gw.IPAddressType
-	addresses := make([]gw.GatewayAddress, 0, len(g.Addresses))
-	for _, address := range g.Addresses {
-		addresses = append(addresses, gw.GatewayAddress{
-			Type:  &ipType,
-			Value: address,
-		})
-	}
-
-	return gw.GatewayStatus{
-		Addresses:  addresses,
-		Conditions: conditions,
-		Listeners:  listenerStatuses,
-	}
-}
-
 func (g *K8sGateway) TrackSync(ctx context.Context, sync func() (bool, error)) error {
 	// we've done all but synced our state, so ensure our deployments are up-to-date
 	if err := g.ensureDeploymentExists(ctx); err != nil {
@@ -407,7 +110,7 @@ func (g *K8sGateway) TrackSync(ctx context.Context, sync func() (bool, error)) e
 		g.GatewayState.Status.InSync = GatewayInSyncStatus{}
 	}
 
-	status := g.Status()
+	status := g.GetStatus(g.Gateway)
 	if !gatewayStatusEqual(status, g.Gateway.Status) {
 		g.Gateway.Status = status
 		if g.logger.IsTrace() {
