--- internal/k8s/reconciler/gateway_state.go
@@ -1,9 +1,0 @@
-package reconciler
-
-// GatewayState holds ephemeral state for gateways
-type GatewayState struct {
-	Status       GatewayStatus
-	PodReady     bool
-	ServiceReady bool
-	Addresses    []string
-}
