+++ internal/k8s/reconciler/manager.go
@@ -45,9 +45,10 @@ type GatewayReconcileManager struct {
 	logger hclog.Logger
 	client gatewayclient.Client
 
-	store          store.Store
-	gatewayClasses *K8sGatewayClasses
-	Factory        *Factory
+	store            store.Store
+	gatewayClasses   *K8sGatewayClasses
+	Factory          *Factory
+	GatewayValidator *GatewayValidator
 
 	consulNamespaceMapper common.ConsulNamespaceMapper
 
@@ -83,6 +84,7 @@ func NewReconcileManager(config ManagerConfig) *GatewayReconcileManager {
 			Client:         config.Client,
 			Resolver:       service.NewBackendResolver(config.Logger, config.ConsulNamespaceMapper, config.Client, config.Consul),
 		}),
+		GatewayValidator:      NewGatewayValidator(config.Client),
 		gatewayClasses:        NewK8sGatewayClasses(config.Logger.Named("gatewayclasses"), config.Client),
 		namespaceMap:          make(map[types.NamespacedName]string),
 		consulNamespaceMapper: config.ConsulNamespaceMapper,
@@ -169,12 +171,12 @@ func (m *GatewayReconcileManager) UpsertGateway(ctx context.Context, g *gw.Gatew
 	})
 
 	// Calling validate outside of the upsert process allows us to re-resolve any
-	// external references and set the statuses accordingly. Since we actually
-	// have other object updates triggering reconciliation loops, this is necessary
-	// prior to dirty-checking on upsert.
-	if err := gateway.Validate(ctx); err != nil {
+	// external references and set the statuses accordingly.
+	state, err := m.GatewayValidator.Validate(ctx, gateway)
+	if err != nil {
 		return err
 	}
+	gateway.SetState(state)
 
 	return m.store.UpsertGateway(ctx, gateway, func(current store.Gateway) bool {
 		if current == nil {
